api/requirements.txt
-i https://pypi.org/simple
flask
mlflow
pandas
scikit-learn
sqlalchemy 
pymysql
python-dotenv